fix: rename `contrib.minijnja` to `contrib.minijinja`
This PR corrects the misspelled module name, while leaving the incorrect namespace available but deprecated.

Closes #2484
